rename all entity with model appended
