+++ src/features/Order/index.ts
@@ -1,2 +1,2 @@
-export * from "./order";
+export * from "./orderModel";
 export * from "./orderStatus";
